--- a/src/spanningtrees/kruskal.jl
+++ b/src/spanningtrees/kruskal.jl
@@ -1,3 +1,8 @@
+"""
+    kruskal_mst(g, distmx=weights(g))
+Return a vector of edges representing the minimum spanning tree of a connected, undirected graph `g` with optional
+distance matrix `distmx` using [Kruskal's algorithm](https://en.wikipedia.org/wiki/Kruskal%27s_algorithm).
+"""
 function kruskal_mst end
 # see https://github.com/mauro3/SimpleTraits.jl/issues/47#issuecomment-327880153 for syntax
 @traitfn function kruskal_mst(
@@ -21,9 +26,7 @@ function kruskal_mst end
         if !DataStructures.in_same_set(connected_vs, e.src, e.dst)
             DataStructures.union!(connected_vs, e.src, e.dst)
             push!(mst, e)
-            if length(mst) >= nv(g) - 1
-                break
-            end
+            (length(mst) >= nv(g) - 1) && break
         end
     end