Improvements due to sorting:

For a complete graph with 1,000 vertices, the run-time improved from 193 ms to 28 ms.
For a path graph with 10,000 vertices, the run-time improved from 3.88 ms to 1.15 ms.

I removed the disjoint set data structure I had created and replaced it with the Inbuilt Disjoint Set. because the improvement in performance was little (28ms to to 26ms)