@sbromberger 2018-06-01
@SohamTamba - thanks for this! 
@somil55 - for your review / decision, please. If you say OK, then I'll be happy to merge.